# util: Improve performance of strings.SplitString